Refactor pending intent builder methods
