+++ app/src/main/java/ch/epfl/sweng/eventmanager/notifications/JoinedEventFeedbackStrategy.java
@@ -30,6 +30,6 @@ public class JoinedEventFeedbackStrategy extends NotificationStrategy<Event> {
 
     private Notification getNotificationFromEvent(Event event) {
         //TODO make the notification on click action send the user to the feedback fragment
-        return NotificationBuilder.getNotificationFromItem(context, titleText + event.getName(), "", event.getId());
+        return NotificationBuilder.getFeedbackNotificationFrom(context, titleText + event.getName(), "", event.getId());
     }
 }
