+++ app/src/main/java/ch/epfl/sweng/eventmanager/notifications/NotificationBuilder.java
@@ -4,12 +4,10 @@ import android.app.Notification;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
-import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.core.app.NotificationCompat;
 import ch.epfl.sweng.eventmanager.R;
 import ch.epfl.sweng.eventmanager.ui.event.interaction.EventShowcaseActivity;
-import ch.epfl.sweng.eventmanager.ui.event.interaction.fragments.EventFeedbackFragment;
 import ch.epfl.sweng.eventmanager.ui.event.selection.EventPickingActivity;
 
 /**
@@ -38,7 +36,14 @@ class NotificationBuilder {
         return PendingIntent.getActivity(context, 0, intent, 0);
     }
 
-    static Notification getNotificationFromItem(@NonNull Context context, String title, String description) {
+    private static PendingIntent toEventMainFragment(Context context, int eventId) {
+        Intent intent = new Intent(context, EventShowcaseActivity.class);
+        intent.putExtra(EventPickingActivity.SELECTED_EVENT_ID, eventId);
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+        return PendingIntent.getActivity(context, 0, intent, 0);
+    }
+
+    static Notification getNotificationFrom(@NonNull Context context, String title, String description) {
         // Set up on tap action TODO send to event my_schedule fragment
         PendingIntent pendingIntent = toEventPickingActivity(context);
 
@@ -46,9 +51,16 @@ class NotificationBuilder {
         return getNotification(context, title, description, pendingIntent);
     }
 
-    static Notification getNotificationFromItem(@NonNull Context context, String title, String description, int eventId) {
-        // Set up on tap action TODO send to event my_schedule fragment
+    static Notification getFeedbackNotificationFrom(@NonNull Context context, String title, String description, int eventId) {
         PendingIntent pendingIntent = toEventFeedbackFragment(context, eventId);
+
+        // Create a notification
+        return getNotification(context, title, description, pendingIntent);
+    }
+
+    static Notification getFirebaseNotificationFrom(@NonNull Context context, String title, String description, int eventId) {
+        // todo make it send to event showcase activity / bug with event feedback
+        PendingIntent pendingIntent = toEventMainFragment(context, eventId);
 
         // Create a notification
         return getNotification(context, title, description, pendingIntent);
